Create separate annotations package for test util annotations
