+++ animatedarchitecture-core/src/test/java/nl/pim16aap2/animatedarchitecture/core/commands/BaseCommandTest.java
@@ -14,8 +14,8 @@ import nl.pim16aap2.animatedarchitecture.core.structures.StructureAttribute;
 import nl.pim16aap2.animatedarchitecture.core.structures.StructureOwner;
 import nl.pim16aap2.animatedarchitecture.core.structures.retriever.StructureRetriever;
 import nl.pim16aap2.animatedarchitecture.core.structures.retriever.StructureRetrieverFactory;
+import nl.pim16aap2.testing.annotations.WithLogCapture;
 import nl.pim16aap2.testing.assertions.AssertionBuilder;
-import nl.pim16aap2.testing.logging.WithLogCapture;
 import nl.pim16aap2.util.exceptions.ContextualOperationException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
