+++ animatedarchitecture-core/src/test/java/nl/pim16aap2/animatedarchitecture/core/extensions/StructureTypeLoaderTest.java
@@ -5,8 +5,8 @@ import com.google.common.jimfs.Jimfs;
 import nl.altindag.log.LogCaptor;
 import nl.pim16aap2.animatedarchitecture.core.api.NamespacedKey;
 import nl.pim16aap2.animatedarchitecture.core.util.Constants;
+import nl.pim16aap2.testing.annotations.WithLogCapture;
 import nl.pim16aap2.testing.assertions.AssertionBuilder;
-import nl.pim16aap2.testing.logging.WithLogCapture;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
