+++ animatedarchitecture-core/src/test/java/nl/pim16aap2/animatedarchitecture/core/structures/StructureAnimationRequestTest.java
@@ -8,7 +8,7 @@ import nl.pim16aap2.animatedarchitecture.core.events.StructureActionCause;
 import nl.pim16aap2.animatedarchitecture.core.events.StructureActionType;
 import nl.pim16aap2.animatedarchitecture.core.structures.properties.Property;
 import nl.pim16aap2.testing.AssistedFactoryMocker;
-import nl.pim16aap2.testing.logging.WithLogCapture;
+import nl.pim16aap2.testing.annotations.WithLogCapture;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
