+++ utilities/test-util/src/main/java/nl/pim16aap2/testing/annotations/WithLogCapture.java
@@ -1,6 +1,7 @@
-package nl.pim16aap2.testing.logging;
+package nl.pim16aap2.testing.annotations;
 
 
+import nl.pim16aap2.testing.logging.LogCaptorExtension;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import java.lang.annotation.ElementType;
