+++ utilities/test-util/src/main/java/nl/pim16aap2/testing/annotations/package-info.java
@@ -1,0 +1,10 @@
+/**
+ * This package contains a set of annotations that are used in tests.
+ * <p>
+ * For example, the {@link nl.pim16aap2.testing.annotations.WithLogCapture} annotation is used to capture logs during
+ * tests.
+ */
+@NonNullByDefault
+package nl.pim16aap2.testing.annotations;
+
+import org.eclipse.jdt.annotation.NonNullByDefault;
